requirements.txt
requests
python-dotenv
yt-dlp
scipy
numpy